Just_A_Suggestion_Game/requirements.txt
setuptools==68.1.2
wheel==0.42.0
fastapi
uvicorn
google-genai
python-dotenv
pydantic
pillow
